refactor(#8): persistence의 엔티티들을 persistence/entity 패키지로 위치 변경
