+++ src/main/java/com/aetheri/infrastructure/persistence/entity/package-info.java
@@ -3,4 +3,4 @@
  * out-port 구현.
  * R2DBC를 사용한 실제 DB 접근 컴포넌트 패키지
  * */
-package com.aetheri.infrastructure.persistence;
+package com.aetheri.infrastructure.persistence.entity;
